pyproject.toml
[project]
name = "aurawell-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "aiosqlite==0.21.0",
    "alembic==1.16.4",
    "alibabacloud-credentials==1.0.2",
    "alibabacloud-credentials-api==1.0.0",
    "alibabacloud-endpoint-util==0.0.4",
    "alibabacloud-fc20230330==4.4.0",
    "alibabacloud-gateway-spi==0.0.3",
    "alibabacloud-openapi-util==0.2.2",
    "alibabacloud-tea==0.4.3",
    "alibabacloud-tea-openapi==0.3.16",
    "alibabacloud-tea-util==0.3.13",
    "alibabacloud-tea-xml==0.0.3",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "apscheduler==3.11.0",
    "attrs==25.3.0",
    "bcrypt==4.3.0",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cryptography==45.0.6",
    "dashvector==1.0.6",
    "distro==1.9.0",
    "ecdsa==0.19.1",
    "fastapi==0.116.1",
    "filelock==3.19.1",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "greenlet>=3.2.4",
    "grpcio==1.74.0",
    "h11==0.16.0",
    "hf-xet==1.1.8",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.34.4",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.4.1",
    "langchain==0.3.27",
    "langchain-core==0.3.75",
    "langchain-openai==0.3.32",
    "langchain-text-splitters==0.3.9",
    "langdetect==1.0.9",
    "langsmith==0.4.19",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "mcp==1.13.1",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "networkx==3.5",
    "numpy==2.3.2",
    "openai==1.102.0",
    "orjson==3.11.3",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "passlib==1.7.4",
    "pluggy==1.6.0",
    "propcache==0.3.2",
    "protobuf==3.20.3",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pysocks==1.7.1",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "redis==6.4.0",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.27.0",
    "rsa==4.9.1",
    "safetensors==0.6.2",
    "selenium==4.35.0",
    "sentencepiece==0.2.1",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sqlalchemy==2.0.43",
    "sse-starlette==3.0.2",
    "starlette==0.47.3",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tiktoken==0.11.0",
    "tokenizers==0.21.4",
    "torch==2.8.0",
    "tqdm==4.67.1",
    "transformers==4.55.4",
    "trio==0.30.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzlocal==5.3.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "webdriver-manager==4.0.2",
    "websocket-client==1.8.0",
    "wsproto==1.2.0",
    "yarl==1.20.1",
    "zstandard==0.24.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
    "--disable-warnings",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "slow: Slow running tests",
    "ai_model: Tests requiring AI model APIs",
    "database: Tests requiring database",
    "redis: Tests requiring Redis",
    "selenium: Tests using Selenium WebDriver",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/aurawell"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["aurawell"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "migrations/",
    "tests/",
]

[dependency-groups]
dev = [
    "pyyaml>=6.0.2",
]
